fix: color 상수 요구사항에 맞게 변경
